Move currently unused code back to root postprocessing_h5py folder
